feat(backend): make the bomb radius 30% bigger
bombRings goes from 8 to 10.4 tile spacings (~230 to ~390 tiles inland).
The field is now a float64, because 10.4 is not a whole number and 10
would give only +25%.

+++ apps/backend/internal/planet/module.go
@@ -263,7 +263,7 @@ func bombRulesOf(config bonus.Config, geography *clicks.Geography) drop_bomb.Rul
 	spacing := geography.Spacing()
 
 	return drop_bomb.Rules{
-		Radius: float64(config.Rings()) * spacing,
+		Radius: config.Rings() * spacing,
 		// Within a tile of the nearest tile is land; further out is the sea.
 		Reach: spacing,
 	}
